+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
-- feat: Make `Table` skip empty thead and tfoot sections
+- feat: Make `Table` skip empty thead and tfoot sections and relax the types
 - feat(ts): Export types `TableCellData` & `TableCellMeta`
 - feat(ts): Add `PropsOf<ComponentType>` utility type
 - fix: `Table`'s cell-level settings should override "cols" meta data